requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-whois>=0.9.4
schedule>=1.2.0
rich>=13.0.0
python-dotenv>=1.0.0
